requirements.txt
torch
torchvision
gradio
numpy==1.26.4
Pillow
